Fix Apache case sensitivity issue
The localhost environment determines that "/folder/file/" and "/Folder/File/" are the same, but on an actual Apache (Linux) server I have found it will return a 404 for the actual case. I have now removed the case correction (implemented to make URLs pretty) but at least it works, if all folder and file names are lowercase this won't effect anything anyway.